requirements.txt
requests>=2.31.0
schedule>=1.2.0
pydantic>=2.4.0
python-dotenv>=1.0.0

transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.2
numpy>=1.24.0

pinecone-client>=5.0.0

plyer>=2.1.0
